animation is ready (need to finish interpolate method)

+++ src/Constructor/Control2D.java
@@ -4,5 +4,5 @@ import javafx.geometry.Point2D;
 
 public interface Control2D {
 
-    void move(Point2D velocity, int speed);
+    void move(Point2D velocity, double speed);
 }
